Title: 사각형 채우기 3, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 250809/사각형 채우기 3/rectangle-fill-3.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int n;
-int D[1001];
+long long D[1001];
 
 int main() {
     cin >> n;
